refactor(policy): move route filtering to policy/adapters
Spec 1.2 complete: Move route filtering functions from
model-routing-helpers.ts to policy/adapters/route-filtering.ts

- Create policy/adapters/route-filtering.ts with:
  - catalogHasDynamicOllamaModel
  - openAiProviderUsesCodexSurface
  - providerRouteSupportsModel
  - filterProviderRoutesByModelSupport
  - filterProviderRoutesByCatalogAvailability
  - shouldRejectModelFromProviderCatalog

- Create catalog-resolution.ts with remaining catalog functions:
  - resolvableConcreteModelIds
  - resolvableConcreteModelIdsForProviders

- Update exports in policy/adapters/index.ts
- Update imports in routes/chat.ts, routes/responses.ts
- Move tests to tests/policy/route-filtering.test.ts
- Update test runner to support subdirectories

All 579 tests pass. No behavior changes.
